plik,plikd: init at 1.3.1
The client and the servers are separated so that a simple user just gets
the necessary binary.

Currently the server frontend has a very old build, I could not build
this asset myself, so for the moment I simply extracted it from the
binary release of the project.

Once this build procedure will have been updated I will transition to a
full build